Add projects watcher. Send notifcations to watchers.

+++ app/controllers/watchers_controller.rb
@@ -4,49 +4,70 @@
 # File: watchers_controller.rb
 
 class WatchersController < ApplicationController
-  before_filter :find_watcher, :only => [:destroy]
+  before_filter :find_watcher, :only => [:create, :destroy]
   before_filter :check_permission, :only => [:destroy, :create]
 
   def create
-    @watcher = Watcher.new(watcher_params)
-    @watcher.author = User.current
+    if @watcher
+      success = @watcher.destroy
+    else
+      new_watcher
+      success = @watcher.save
+    end
     respond_to do |format|
-      if @watcher.save
-        format.js { respond_to_js :response_header => :success, :response_content => t(:successful_creation) }
+      if success
+        format.js { respond_to_js :response_header => :success, :response_content => t(:successful_deletion) }
       else
-        format.js { respond_to_js action: 'do_nothing', :response_header => :failure, :response_content => "#{t(:failure_creation)} : #{@watcher.errors.full_messages.join(', ')}" }
+        format.js { respond_to_js action: 'do_nothing', :response_header => :failure, :response_content => "#{t(:failure_deletion)} : #{@watcher.errors.full_messages.join(', ')}" }
       end
     end
   end
 
   def destroy
-    if @watcher.user_id.eql? User.current.id
+    if @watcher && @watcher.user_id.eql?(User.current.id)
       @model = @watcher.watchable
       respond_to do |format|
         if @watcher.destroy
-          format.js { respond_to_js :response_header => :success, :response_content => t(:successful_deletion) }
+          format.js { respond_to_js action: 'destroy', :response_header => :success, :response_content => t(:successful_deletion) }
         else
           format.js { respond_to_js action: 'do_nothing', :response_header => :failure, :response_content => "#{t(:failure_deletion)} : #{@watcher.errors.full_messages.join(', ')}" }
+        end
+      end
+    else
+      new_watcher
+      @watcher.is_unwatch = true
+      respond_to do |format|
+        if @watcher.save
+          @model = @watcher.watchable
+          format.js { respond_to_js :response_header => :success, :response_content => t(:successful_creation) }
+        else
+          format.js { respond_to_js action: 'do_nothing', :response_header => :failure, :response_content => "#{t(:successful_creation)} : #{@watcher.errors.full_messages.join(', ')}" }
         end
       end
     end
   end
 
   private
-  def watcher_params
-    params.require(:watcher).permit(Watcher.permit_attributes)
-  end
 
   def find_watcher
-    @watcher = Watcher.find_by_id(params[:id])
+    @watcher = Watcher.find_by_watchable_id_and_watchable_type_and_user_id(params[:watchable_id], params[:watchable_type], User.current.id)
   end
 
   def check_permission
-    controller = @watcher ? @watcher.watchable_type.pluralize : params[:watcher][:watchable_type].pluralize
-    if User.current.allowed_to?('watch', controller, @project)
+    controller = @watcher ? @watcher.watchable_type.pluralize : params[:watchable_type].pluralize
+    project = params[:watchable_type].eql?('Project') ? nil : @project
+    if User.current.allowed_to?('watch', controller, project)
       true
     else
       render_403
     end
+  end
+
+  def new_watcher
+    @watcher = Watcher.new
+    @watcher.watchable_id = params[:watchable_id]
+    @watcher.watchable_type = params[:watchable_type]
+    @watcher.author = User.current
+    @watcher.project = @project
   end
 end
